chore: move test files to the test folder
